[refacto] region filter + kanban list
